Use Travis CI to test various compilers
Currently we test clang-7 and gcc-7 with their corresponding C++
compilers.  This is because these are the versions available on
Travis-CI that support C++17. We install a newer version of CMake that
will eventually be required to test the python bindings.  We disable the
various plugins because they require additional dependencies.

No changes were made to libpressio itself.