Share initialization between sz plugins
Previously sz_threadsafe did not initialize itself to default
parameters.  Now both sz and sz_threadsafe must grab a sz_init_handle
which is responsible to call SZ_Init and finalize.  Additionally,
if the plugin wants to write to confparams_cpr, it must grab an
exclusive lock to indicate it is configuring sz.  If a plugin wants to
read from confparams_cpr, it must grab a shared lock.